# packages/api-contract/openapi.yaml
openapi: 3.0.3
info:
  title: i-um API
  version: 0.1.0
  description: API contract for the i-um mobile app.
servers:
  - url: http://localhost:8080
    description: Local development
...
paths:
...
  /trips:
    get:
      operationId: listTrips
      summary: List my trips
      description: Returns trips where the authenticated user is an Owner or Member participant.
      tags:
        - Trips
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Participated trips for the authenticated user.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListTripsResponse'
        '401':
          description: Unauthorized.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Unexpected server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      operationId: createTrip
      summary: Create a trip
      description: Creates a Trip and the authenticated user's Owner participant.
      tags:
        - Trips
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTripRequest'
      responses:
        '201':
          description: Trip created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateTripResponse'
        '400':
          description: Validation error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Unexpected server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /trips/{tripId}:
    get:
      operationId: getTripDetail
      summary: Get trip detail
      description: Returns basic trip detail for an authenticated trip participant.
      tags:
        - Trips
      security:
        - bearerAuth: []
      parameters:
        - name: tripId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Trip detail.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetTripDetailResponse'
        '400':
          description: Validation error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Trip not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Unexpected server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    patch:
      operationId: updateTrip
      summary: Update trip basic information
      description: Updates basic trip metadata for the authenticated trip owner. Only changed fields need to be sent.
      tags:
        - Trips
      security:
        - bearerAuth: []
      parameters:
        - name: tripId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTripRequest'
      responses:
        '200':
          description: Trip updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UpdateTripResponse'
        '400':
          description: Validation error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Trip not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Unexpected server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      operationId: deleteTrip
      summary: Delete a trip
      description: Hard-deletes a trip for all participants when requested by the authenticated trip owner.
      tags:
        - Trips
      security:
        - bearerAuth: []
      parameters:
        - name: tripId
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Trip deleted.
        '400':
          description: Validation error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Trip not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Unexpected server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /trips/{tripId}/invites:
    post:
      operationId: createTripInvite
      summary: Create or retrieve the current trip invite link
      description: Creates a new invite link for a trip owner, or returns the existing unexpired current invite link.
      tags:
        - Trips
      security:
        - bearerAuth: []
      parameters:
        - name: tripId
          in: path
          required: true
          schema:
            type: string
      responses:
        '201':
          description: New invite link created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateTripInviteResponse'
        '200':
          description: Existing active invite link returned.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateTripInviteResponse'
        '400':
          description: Validation error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Trip not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Invite token conflict after retry.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Unexpected server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /invites/{token}/accept:
    post:
      operationId: acceptTripInvite
      summary: Accept a trip invite link
      description: Accepts a reusable active trip invite for the authenticated user. Existing participants receive an idempotent success response.
      tags:
        - Trips
      security:
        - bearerAuth: []
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
            minLength: 32
            maxLength: 128
            pattern: '^[A-Za-z0-9_-]+$'
      responses:
        '200':
          description: Invite accepted or already accepted.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AcceptTripInviteResponse'
        '400':
          description: Validation error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Invite token not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '410':
          description: Invite token expired or deactivated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Unexpected server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /trips/{tripId}/participants:
    get:
      operationId: listTripParticipants
      summary: List trip participants
      description: Returns the current accepted participants for a trip. Only authenticated current trip participants can access the list.
      tags:
        - Trips
      security:
        - bearerAuth: []
      parameters:
        - name: tripId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Current trip participants.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListTripParticipantsResponse'
        '400':
          description: Validation error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Trip not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Unexpected server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /trips/{tripId}/participants/{participantId}:
    delete:
      operationId: removeTripParticipant
      summary: Remove a trip participant
      description: Removes a member participant from a trip when requested by the authenticated trip owner. Removed members immediately lose trip access and may rejoin through a valid invite link.
      tags:
        - Trips
      security:
        - bearerAuth: []
      parameters:
        - name: tripId
          in: path
          required: true
          schema:
            type: string
        - name: participantId
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Participant removed.
        '400':
          description: Validation error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Trip or removable participant not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Unexpected server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /trips/{tripId}/days/{date}/lodging-place:
    put:
      operationId: setDayLodgingPlace
      summary: Set a trip day lodging place
      description: Stores or replaces the selected Day lodging target for an authenticated trip participant.
      tags:
        - Trips
      security:
        - bearerAuth: []
      parameters:
        - name: tripId
          in: path
          required: true
          schema:
            type: string
        - name: date
          in: path
          required: true
          schema:
            type: string
            format: date
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SetDayLodgingPlaceRequest'
      responses:
        '200':
          description: Day lodging place saved.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SetDayLodgingPlaceResponse'
        '400':
          description: Validation error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Trip, virtual day, or trip place not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Unexpected server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      operationId: clearDayLodgingPlace
      summary: Clear a trip day lodging place
      description: Removes the selected Day lodging target for an authenticated trip participant. Clearing an empty Day lodging selection is a 204 no-op.
      tags:
        - Trips
      security:
        - bearerAuth: []
      parameters:
        - name: tripId
          in: path
          required: true
          schema:
            type: string
        - name: date
          in: path
          required: true
          schema:
            type: string
            format: date
      responses:
        '204':
          description: Day lodging place cleared or already empty.
        '400':
          description: Validation error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Trip or virtual day not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Unexpected server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /trips/{tripId}/days/{date}/itinerary:
    get:
      operationId: getDayItinerary
      summary: Get a trip day itinerary
      description: Returns the ordered read-only itinerary items for a selected virtual trip day.
      tags:
        - Trips
      security:
        - bearerAuth: []
      parameters:
        - name: tripId
          in: path
          required: true
          schema:
            type: string
        - name: date
          in: path
          required: true
          schema:
            type: string
            format: date
      responses:
        '200':
          description: Day itinerary.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetDayItineraryResponse'
        '400':
          description: Validation error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Trip or virtual day not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Unexpected server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    CreateTripRequest:
      type: object
      required:
        - name
        - startDate
        - endDate
        - defaultCurrency
      additionalProperties: false
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 80
          example: 오사카 3박 4일
        startDate:
          type: string
          format: date
          description: Start date in YYYY-MM-DD format. Must be today or later.
          example: '2026-07-10'
        endDate:
          type: string
          format: date
          description: End date in YYYY-MM-DD format. Must be today or later and not before startDate.
          example: '2026-07-13'
        defaultCurrency:
          $ref: '#/components/schemas/SupportedCurrency'
...
    UpdateTripResponse:
      type: object
      required:
        - trip
      properties:
        trip:
          $ref: '#/components/schemas/Trip'
...
    ListTripsResponse:
      type: object
      required:
        - trips
      properties:
        trips:
          type: array
          items:
            $ref: '#/components/schemas/TripListItem'
...
    ListTripParticipantsResponse:
      type: object
      required:
        - participants
      properties:
        participants:
          type: array
          items:
            $ref: '#/components/schemas/TripParticipantListItem'
...
    CreateTripResponse:
      type: object
      required:
        - trip
        - ownerParticipant
      properties:
        trip:
          $ref: '#/components/schemas/Trip'
        ownerParticipant:
          $ref: '#/components/schemas/TripParticipant'
...
    GetTripDetailResponse:
      type: object
      required:
        - trip
        - participantSummary
        - days
      properties:
        trip:
          $ref: '#/components/schemas/Trip'
        participantSummary:
          $ref: '#/components/schemas/TripParticipantSummary'
        days:
          type: array
          items:
            $ref: '#/components/schemas/TripDay'
...
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: object
          required:
            - code
            - message
            - details
          properties:
            code:
              type: string
              example: INTERNAL_ERROR
            message:
              type: string
              example: internal server error
            details:
              type: array
              items:
                type: object
                additionalProperties: true